requirements.txt.txt
streamlit==1.39.0
pandas==2.1.0
scikit-learn==1.3.0
scikit-surprise==1.1.3
numpy==1.24.3
sentence-transformers==3.0.1